index rework

+++ program.py
@@ -49,28 +49,30 @@ def logout():
         session.pop('logged', None)
         return redirect(url_for('login'))
 
-
-
-    
-
 @app.route('/index')
 def index():
     if 'logged' in session:
-        return render_template('index.html')
+        data = postgres.indexData()
+        return render_template('index.html', data = data)
     else:
         return redirect(url_for('login'))
 
 @app.route('/users', methods= ['GET','POST'])
 def users():
     if 'logged' in session:
-        if request.method == 'POST':
-            users = forms.searched(request)
-            users, page, per_page,pagination,paglen = func.paginateResults(users,5)
-            return render_template('pages/users.html', users=users, page=page,per_page=per_page,pagination=pagination,paglen=paglen)
+        if session['admin']:
+            if request.method == 'POST':
+                users = forms.searched(request)
+                users, page, per_page,pagination,paglen = func.paginateResults(users,5)
+                return render_template('pages/users.html', users=users, page=page,per_page=per_page,pagination=pagination,paglen=paglen)
+            else:
+                users = postgres.findUsers()
+                users, page, per_page,pagination,paglen = func.paginateResults(users,5)
+                return render_template('pages/users.html', users=users, page=page,per_page=per_page,pagination=pagination,paglen=paglen)
         else:
-            users = postgres.findUsers()
-            users, page, per_page,pagination,paglen = func.paginateResults(users,5)
-            return render_template('pages/users.html', users=users, page=page,per_page=per_page,pagination=pagination,paglen=paglen)
+            data = postgres.indexData()
+            flash('Sorry, you don\'t have the permission to view this page')
+            return redirect(url_for('index', data=data))
     else:
         return redirect(url_for('login'))
 
@@ -78,55 +80,88 @@ def users():
 @app.route('/new_user', methods=['GET', 'POST'])
 def new_user():
     if 'logged' in session:
-        if request.method == 'POST' and len(request.form) > 0:
-            forms.addUser(request)
-            return redirect(url_for('users'))
-        return render_template('forms/new_user.html')
+        if session['admin']:
+            if request.method == 'POST' and len(request.form) > 0:
+                forms.addUser(request)
+                return redirect(url_for('users'))
+            return render_template('forms/new_user.html')
+        else:
+            data = postgres.indexData()
+            flash('Sorry, you don\'t have the permission to view this page')
+            return redirect(url_for('index', data=data))
     else:
         return redirect(url_for('login'))
 
 @app.route('/books', methods= ['GET','POST'])
 def books():
     if 'logged' in session:
-        if request.method == 'POST':
-            books = forms.searched(request)
-            books, page, per_page,pagination,paglen = func.paginateResults(books,10)
-            return render_template('pages/books.html', books=books, page=page,per_page=per_page,pagination=pagination,paglen=paglen)
+        if session['admin']:
+            if request.method == 'POST':
+                books = forms.searched(request)
+                books, page, per_page,pagination,paglen = func.paginateResults(books,10)
+                return render_template('pages/books.html', books=books, page=page,per_page=per_page,pagination=pagination,paglen=paglen)
+            else:
+                books = postgres.findBooks()
+                books, page, per_page,pagination,paglen = func.paginateResults(books,10)
+                return render_template('pages/books.html', books=books, page=page,per_page=per_page,pagination=pagination,paglen=paglen)
         else:
-            books = postgres.findBooks()
-            books, page, per_page,pagination,paglen = func.paginateResults(books,10)
-            return render_template('pages/books.html', books=books, page=page,per_page=per_page,pagination=pagination,paglen=paglen)
+            data = postgres.indexData()
+            flash('Sorry, you don\'t have the permission to view this page')
+            return redirect(url_for('index', data=data))
     else:
         return redirect(url_for('login'))
 
 @app.route('/new_book', methods=['GET','POST'])
 def new_book():
     if 'logged' in session:
-        if request.method == 'POST' and len(request.form) > 0:
-            forms.addBook(request)
-            return redirect(url_for('books'))
-        return render_template('forms/new_book.html')
+        if session['admin']:
+            if request.method == 'POST' and len(request.form) > 0:
+                forms.addBook(request)
+                return redirect(url_for('books'))
+            return render_template('forms/new_book.html')
+        else:
+            data = postgres.indexData()
+            flash('Sorry, you don\'t have the permission to view this page')
+            return redirect(url_for('index', data=data))
     else:
         return redirect(url_for('login'))
 
 
-@app.route('/issued_books')
+@app.route('/issued_books', methods=['GET', 'POST'])
 def issued_books():
     if 'logged' in session:
-        ibooks = postgres.findRentals()
-        return render_template('pages/issued_books.html', ibooks=ibooks)
+        if session['admin']:
+            if request.method == 'POST':
+                ibooks = forms.searched(request)
+                ibooks, page, per_page,pagination,paglen = func.paginateResults(ibooks,10)
+                return render_template('pages/issued_books.html', ibooks=ibooks, page=page,per_page=per_page,pagination=pagination,paglen=paglen)
+                
+            else:
+                ibooks = postgres.findRentals()
+                ibooks, page, per_page,pagination,paglen = func.paginateResults(ibooks,10)
+                return render_template('pages/issued_books.html', ibooks=ibooks, page=page,per_page=per_page,pagination=pagination,paglen=paglen)
+        else:
+            data = postgres.indexData()
+            flash('Sorry, you don\'t have the permission to view this page')
+            return redirect(url_for('index', data=data))
     else:
         return redirect(url_for('login'))
 
+# Add from issued books
 @app.route('/issue_book', methods=['GET','POST'])
 def issue_book():
     if 'logged' in session:
-        bid = postgres.findBooks()
-        uid = postgres.findUsers()
-        if request.method == 'POST' and len(request.form) > 0:
-            forms.addRental(request)
-            return redirect(url_for('issued_books'))
-        return render_template('forms/issue_book.html', bid=bid, uid=uid)
+        if session['admin']:
+            bid = postgres.findBooks()
+            uid = postgres.findUsers()
+            if request.method == 'POST' and len(request.form) > 0:
+                forms.addRental(request)
+                return redirect(url_for('issued_books'))
+            return render_template('forms/issue_book.html', bid=bid, uid=uid)
+        else:
+            data = postgres.indexData()
+            flash('Sorry, you don\'t have the permission to view this page')
+            return redirect(url_for('index', data=data))
     else:
         return redirect(url_for('login'))
 
@@ -134,11 +169,16 @@ def issue_book():
 @app.route('/return_book', methods=['GET','POST'])
 def retrun_books():
     if 'logged' in session:
-        rid = postgres.findReturnBooks()
-        if request.method == 'POST' and len(request.form) > 0:
-            forms.returnRental(request)
-            return redirect(url_for('issued_books'))
-        return render_template('forms/return_book.html', rid=rid)
+        if session['admin']:
+            rid = postgres.findReturnBooks()
+            if request.method == 'POST' and len(request.form) > 0:
+                forms.returnRental(request)
+                return redirect(url_for('issued_books'))
+            return render_template('forms/return_book.html', rid=rid)
+        else:
+            data = postgres.indexData()
+            flash('Sorry, you don\'t have the permission to view this page')
+            return redirect(url_for('index', data=data))
     else:
         return redirect(url_for('login'))
 
@@ -146,10 +186,9 @@ def retrun_books():
 @app.route('/edit_user/<int:uid>', methods=['GET', 'POST'])
 def edit_user(uid):
     if 'logged' in session:
-        u = postgres.findUserById(uid)
-
         ## Code for restricting access
         if session['admin']:
+            u = postgres.findUserById(uid)
             if request.method == 'POST' and len(request.form) > 0:
                 if (forms.editUser(request, u)):
                     flash('User has been edited succesfully')
@@ -160,8 +199,9 @@ def edit_user(uid):
             else:
                 return render_template('forms/edits/edit_user.html',user=u)
         else:
+            data = postgres.indexData()
             flash('Sorry, you don\'t have the permission to view this page')
-            return redirect(url_for('index'))
+            return redirect(url_for('index', data=data))
     else:
         return redirect(url_for('login'))
 
@@ -180,35 +220,44 @@ def edit_book(bid):
             else:
                 return render_template('forms/edits/edit_book.html',book=b)
         else:
-            flash('Sorry, you don\'t have the permission to view this page')
-            return redirect(url_for('index'))
+            flash('Sorry, you don\'t have the permission to edit books')
+            return redirect(url_for('books'))
     else:
         return redirect(url_for('login'))
 
 @app.route('/delete_book/<int:bid>')
 def delete_book(bid):
     if 'logged' in session:
-        if(forms.deleteBook(bid)):
-            flash("The book has been deleted succesfully")
-            return redirect(url_for('books'))
+        if session['admin']:
+            if(forms.deleteBook(bid)):
+                flash("The book has been deleted succesfully")
+                return redirect(url_for('books'))
+            else:
+                flash("The book has not been deleted")
+                return redirect(url_for('books'))
         else:
-            flash("The book has not been deleted")
-            return redirect(url_for('books'))
+            data = postgres.indexData()
+            flash('Sorry, you don\'t have the permission to view this page')
+            return redirect(url_for('index', data=data))
     else:
         return redirect(url_for('login'))
 
 @app.route('/delete_user/<int:uid>')
 def delete_user(uid):
     if 'logged' in session:
-        if(forms.deleteUser(uid)):
-            flash("The user has been deleted succesfully")
-            return redirect(url_for('users'))
+        if session['admin']:
+            if(forms.deleteUser(uid)):
+                flash("The user has been deleted succesfully")
+                return redirect(url_for('users'))
+            else:
+                flash("The user has not been deleted")
+                return redirect(url_for('users'))
         else:
-            flash("The user has not been deleted")
-            return redirect(url_for('users'))
+            data = postgres.indexData()
+            flash('Sorry, you don\'t have the permission to view this page')
+            return redirect(url_for('index', data=data))
     else:
         return redirect(url_for('login'))
-
